Standard-evaluation version of

This is a S3 generic.

extract_(data, col, into, regex = "([[:alnum:]]+)", remove = TRUE,
  convert = FALSE, ...)

Arguments

data

A data frame.

col

Name of column to split, as string.

into

Names of new variables to create as character vector.

regex

a regular expression used to extract the desired values.

remove

If TRUE, remove input column from output data frame.

convert

If TRUE, will run type.convert with as.is = TRUE on new columns. This is useful if the component columns are integer, numeric or logical.

...

Other arguments passed on to regexec to control how the regular expression is processed.