Fill in missing values.

Fills missing values in using the previous entry. This is useful in the common output format where values are not repeated, they're recorded each time they change.

fill(data, ..., .direction = c("down", "up"))

Arguments

data
A data frame.

...
Specification of columns to fill. Use bare variable names. Select all variables between x and z with x:z, exclude y with -y. For more options, see the select documentation.

.direction
Direction in which to fill missing values. Currently either "down" (the default) or "up".

Details

Missing values are replaced in atomic vectors; NULLs are replaced in list.

See also

fill_ for a version that uses regular evaluation and is suitable for programming with.

Examples

df <- data.frame(Month = 1:12, Year = c(2000, rep(NA, 11)))
df %>% fill(Year)

#>    Month Year
#> 1      1 2000
#> 2      2 2000
#> 3      3 2000
#> 4      4 2000
#> 5      5 2000
#> 6      6 2000
#> 7      7 2000
#> 8      8 2000
#> 9      9 2000
#> 10    10 2000
#> 11    11 2000
#> 12    12 2000