Pivot a data frame from wide to long or long to wide

pivot() is provides rectangular reshaping like gather() and spread(). It differs primarily from existing approaches in tidyr because the details of the reshaping a described by a data frame, spec. For simple cases, you can construct the spec with pivot_spec_long() and pivot_spec_wide(). See details in vignette("pivot")

pivot_spec_long(df, cols, measure = "value", variable = "variable")

pivot_spec_wide(df, cols, measure = value, sep = "_")

pivot(df, spec, na.rm = FALSE, ptypes = NULL)

Arguments

df
A data frame to reshape.

spec
A data frame defining the reshaping specification. Must contain col_name and measure columns that are character vectors.

na.rm
If TRUE, will convert explicit missing values to implicit missing values. Used only when pivotting to long.

.ptype
A named list that optionally override the types of measured columns. Used only when pivotting to long.