Pivot data from long to wide

pivot_wide() "widens" data, increasing the number of columns and decreasing the number of rows. See more details in vignette("pivot"), and see pivot_long() for the inverse transformation.

pivot_wide_spec(df, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_")

pivot_wide(df, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_", values_fill = NULL,
  spec = NULL)

Arguments

df
A data frame to pivot.

names_from, values_from
A pair of arguments describing which column (or columns) to get the name of the output column (name_from), and which column to get the cell values from (values_from).

names_prefix
String added to the start of every variable name. This is particularly useful if names_from is a numeric vector and you want to create syntactic variable names.

names_sep
If names_from contains multiple variables, this will be used to join their values together into a single string to use as a column name.

values_fill
Optionally, a named list specifying what each value should be filled in with when missing.

spec
Alternatively, instead of providing cols (and names_to and values_to) you can parse a specification data frame. This is useful for more complex pivots because it gives you greater control on how metadata stored in the column names turns into columns in the result.

Must be a data frame containing character .name and .value columns.