uncount {tidyr}	R Documentation

"Uncount" a data frame

Description

Performs the opposite operation to dplyr::count(), duplicating rows according to a weighting variable (or expression).

Usage

uncount(data, weights, ..., .remove = TRUE, .id = NULL)

Arguments

data
A data frame, tibble, or grouped tibble.

weights
A vector of weights. Evaluated in the context of data; supports quasiquotation.

...
Additional arguments passed on to methods.

.remove
If TRUE, and weights is the name of a column in data, then this column is removed.

.id
Supply a string to create a new variable which gives a unique identifier for each created row.

Examples

df <- tibble(x = c("a", "b"), n = c(1, 2))
uncount(df, n)
uncount(df, n, .id = "id")

# You can also use constants
uncount(df, 2)

# Or expressions
uncount(df, 2 / n)

[Package tidyr version 1.3.0 Index]