Standard-evaluation version of unnest.

This is a S3 generic.

unnest_(data, unnest_cols, .drop = NA, .id = NULL, .sep = NULL)

Arguments

data
A data frame.

unnest_cols
Name of columns that needs to be unnested.

.drop
Should additional list columns be dropped? By default, unnest will drop them if unnesting the specified columns requires the rows to be duplicated.

.id
Data frame idenfier - if supplied, will create a new column with name .id, giving a unique identifer. This is most useful if the list column is named.

.sep
If non-NULL, the names of unnested data frame columns will combine the name of the original list-col with the names from nested data frame, separated by .sep.