Chop and unchop

Chopping and unchopping preserve the width of a data frame, changing its length. chop() makes df shorter by converting rows within each group into list-columns. unchop() makes df longer by expanding list-columns so that each element of the list-column gets its own row in the output.

chop(data, cols)

unchop(data, cols, keep_empty = FALSE, ptype = NULL)

Arguments

data
A data frame.

cols
Column to chop or unchop (automatically quoted).

This should be a list-column containing generalised vectors (e.g. any mix of NULLs, atomic vector, S3 vectors, a lists, or data frames).

keep_empty
By default, elements of col that have size zero will be ommitted from the output. Setting keep_empty = TRUE will ensure that they're preserved

ptype
Optionally, supply a data frame prototype for the output cols, overriding the default that will be guessed from the combination of individual values.

Details

Generally, unchopping is more useful than chopping because it simplifies a complex data structure, and nest()ing is usually more appropriate that chop()ing` since it better preserves the connections between observations.

Examples

# Chop ==============================================================
df <- tibble(x = c(1, 1, 1, 2, 2, 3), y = 1:6, z = 6:1)
# Note that we get one row of output for each unique combination of
# non-chopped variables
df %>% chop(c(y, z))

#> # A tibble: 3 x 3
#>       x y         z        
#>   <dbl> <list>    <list>   
#> 1     1 <int [3]> <int [3]>
#> 2     2 <int [2]> <int [2]>
#> 3     3 <int [1]> <int [1]>

# cf nest
df %>% nest(c(y, z))

#> Warning: All elements of `...` must be named.
#> Did you want `data = c(y, z)`?

#> # A tibble: 3 x 2
#>       x           data
#>   <dbl> <list<df[,2]>>
#> 1     1        [3 × 2]
#> 2     2        [2 × 2]
#> 3     3        [1 × 2]

# Unchop ============================================================
df <- tibble(x = 1:4, y = list(integer(), 1L, 1:2, 1:3))
df %>% unchop(y)

#> # A tibble: 6 x 2
#>       x     y
#>   <int> <int>
#> 1     2     1
#> 2     3     1
#> 3     3     2
#> 4     4     1
#> 5     4     2
#> 6     4     3

df %>% unchop(y, keep_empty = TRUE)

#> # A tibble: 7 x 2
#>       x     y
#>   <int> <int>
#> 1     1    NA
#> 2     2     1
#> 3     3     1
#> 4     3     2
#> 5     4     1
#> 6     4     2
#> 7     4     3

# Incompatible types -------------------------------------------------
# If the list-col contains types that can not be natively
df <- tibble(x = 1:2, y = list("1", 1:3))
try(df %>% unchop(y))

#> Error : No common type for `list(...)[[1]]$y` <character> and `list(...)[[2]]$y` <integer>.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. ├─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13. │ └─pkgdown::build_site(...)
#>  14. │   └─pkgdown:::build_site_local(...)
#>  15. │     └─pkgdown::build_reference(...)
#>  16. │       └─purrr::map(...)
#>  17. │         └─pkgdown:::.f(.x[[i]], ...)
#>  18. │           └─pkgdown:::data_reference_topic(...)
#>  19. │             ├─pkgdown:::as_data(...)
#>  20. │             └─pkgdown:::as_data.tag_examples(...)
#>  21. │               └─purrr::pmap_chr(...)
#>  22. │                 └─pkgdown:::.f(...)
#>  23. │                   ├─withr::with_options(...)
#>  24. │                   │ └─base::force(code)
#>  25. │                   └─evaluate::evaluate(code, env, new_device = TRUE)
#>  26. │                     └─evaluate:::evaluate_call(...)
#>  27. │                       ├─evaluate:::timing_fn(...)
#>  28. │                       ├─evaluate:::handle(...)
#>  29. │                       │ └─base::try(f, silent = TRUE)
#>  30. │                       │   └─base::tryCatch(...)
#>  31. │                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  32. │                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  33. │                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  34. │                       ├─base::withCallingHandlers(...)
#>  35. │                       ├─base::withVisible(eval(expr, envir, enclos))
#>  36. │                       └─base::eval(expr, envir, enclos)
#>  37. │                         └─base::eval(expr, envir, enclos)
#>  38. │                           ├─base::try(df %>% unchop(y))
#>  39. │                           │ └─base::tryCatch(...)
#>  40. │                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  41. │                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  42. │                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  43. │                           └─df %>% unchop(y)
#>  44. │                             ├─base::withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#>  45. │                             └─base::eval(quote(`_fseq`(`_lhs`)), env, env)
#>  46. │                               └─base::eval(quote(`_fseq`(`_lhs`)), env, env)
#>  47. │                                 └─global::`_fseq`(`_lhs`)
#>  48. │                                   └─magrittr::freduce(value, `_function_list`)
#>  49. │                                     ├─base::withVisible(function_list[[k]](value))
#>  50. │                                     └─function_list[[k]](value)
#>  51. │                                       └─tidyr::unchop(., y)
#>  52. │                                         └─vctrs::vec_rbind(!!!x, .ptype = ptype) /home/travis/build/tidyverse/tidyr/R/chop.R:93:4
#>  53. │                                           └─vctrs::vec_type_common(!!!tbls, .ptype = .ptype)
#>  54. ├─vctrs:::vec_type2_dispatch(x = x, y = y, x_arg = x_arg, y_arg = y_arg)
#>  55. ├─vctrs::vec_type2.tbl_df(x = x, y = y, x_arg = x_arg, y_arg = y_arg)
#>  56. ├─vctrs:::vec_type2.tbl_df.data.frame(...)
#>  57. │ └─vctrs:::df_col_type2(x, y, ...)
#>  58. │   └─vctrs:::pmap(...)
#>  59. │     ├─base::do.call(...)
#>  60. │     └─base::mapply(...)
#>  61. │       └─(function (x, y, ..., x_arg = "", y_arg = "") ...
#>  62. ├─vctrs:::vec_type2_dispatch(x = x, y = y, x_arg = x_arg, y_arg = y_arg)
#>  63. ├─vctrs::vec_type2.character(x = x, y = y, x_arg = x_arg, y_arg = y_arg)
#>  64. └─vctrs:::vec_type2.character.default(...)
#>  65.   └─vctrs::stop_incompatible_type(x, y, x_arg = x_arg, y_arg = y_arg)
#>  66.     └─vctrs:::stop_incompatible(...)

df %>% unchop(y, ptype = tibble(y = integer()))

#> # A tibble: 4 x 2
#>       x     y
#>   <int> <int>
#> 1     1     1
#> 2     2     1
#> 3     2     2
#> 4     2     3

df %>% unchop(y, ptype = tibble(y = character()))

#> # A tibble: 4 x 2
#>       x y    
#>   <int> <chr>
#> 1     1 1    
#> 2     2 1    
#> 3     2 2    
#> 4     2 3    

df %>% unchop(y, ptype = tibble(y = list()))

#> # A tibble: 4 x 2
#>       x y        
#>   <int> <list>   
#> 1     1 <chr [1]>
#> 2     2 <int [1]>
#> 3     2 <int [1]>
#> 4     2 <int [1]>

# Unchopping data frames -----------------------------------------------------
# Unchopping a list-col of data frames must generate a df-col because
# unchop leaves the column names unchanged
df <- tibble(x = 1:3, y = list(NULL, tibble(x = 1), tibble(y = 1:2)))
df %>% unchop(y)

#> # A tibble: 3 x 2
#>       x   y$x    $y
#>   <int> <dbl> <int>
#> 1     2     1    NA
#> 2     3    NA     1
#> 3     3    NA     2

df %>% unchop(y, keep_empty = FALSE)

#> # A tibble: 3 x 2
#>       x   y$x    $y
#>   <int> <dbl> <int>
#> 1     2     1    NA
#> 2     3    NA     1
#> 3     3    NA     2