Nest repeated values in a list-column of data frames

Nesting is implicitly a summarising operation: you get one row defined by the grouping defined by the non-nested columns. This is useful in conjunction with other summaries that work with whole datasets, most notably models, as described in https://r4ds.had.co.nz/many-models.html.

Learn more in vignette("nesting").

nest(.data, ..., .key = "DEPRECATED")

Arguments

.data
A data frame.

...
Name-variable pairs of the form new_col = c(col1, col2, col3), that describe how you wish to nest existing columns into new columns. The right hand side can be any expression supported by tidyselect.

Deprecated: previously you could write df %>% nest(x, y, z), now you should rewrite this as df %>% nest(data = c(x, y, z)).

.key
Deprecated: No longer needed because of the updated ... syntax.

Grouped data frames

nest() specify the columns to be nested; i.e. the columns that will appear in the inner data frame. Alternatively, you nest() a grouped data frame created by dplyr::group_by(). Here the grouping variables will remain in the outer data frame, and all other variables will go into the inner data frame. The result will be an ungrouped tibble, where each row represents a group in the input.

See also

Examples

df <- tibble(x = c(1, 1, 1, 2, 2, 3), y = 1:6, z = 6:1)
# Note that we get one row of output for each unique combination of
# non-chopped variables
df %>% nest(data = c(y, z))

#> # A tibble: 3 x 2
#>       x             data
#>   <dbl>   <list<df[,2]>>
#> 1     1 1, 2, 3, 6, 5, 4
#> 2     2       4, 5, 3, 2
#> 3     3             6, 1

# cf nest
df %>% chop(c(y, z))

#> # A tibble: 3 x 3
#>       x y         z        
#>   <dbl> <list>    <list>   
#> 1     1 <int [3]> <int [3]>
#> 2     2 <int [2]> <int [2]>
#> 3     3 <int [1]> <int [1]>

iris %>% nest(data = -Species)

#> # A tibble: 3 x 2
#>   Species
#>   <fct>  
#> 1 setosa 
#> 2 versic…
#> 3 virgin…
#> # … with 1 more variable: data <list<df[,4]>>

chickwts %>% nest(data = weight)

#> # A tibble: 6 x 2
#>   feed                                                                      data
#>   <fct>                                                           <list<df[,1]>>
#> 1 horsebean                     179, 160, 136, 227, 217, 168, 108, 124, 143, 140
#> 2 linseed             309, 229, 181, 141, 260, 203, 148, 169, 213, 257, 244, 271
#> 3 soybean   243, 230, 248, 327, 329, 250, 193, 271, 316, 267, 199, 171, 158, 248
#> 4 sunflower           423, 340, 392, 339, 341, 226, 320, 295, 334, 322, 297, 318
#> 5 meatmeal                 325, 257, 303, 315, 380, 153, 263, 242, 206, 344, 258
#> 6 casein              368, 390, 379, 260, 404, 318, 352, 359, 216, 222, 283, 332

# Nesting a grouped data frame nests all variables apart from the group vars
fish_encounters %>%
  dplyr::group_by(fish) %>%
  nest()

#> # A tibble: 19 x 2
#>    fish                                                                data
#>    <fct>                                                     <list<df[,2]>>
#>  1 4842  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#>  2 4843  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#>  3 4844  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#>  4 4845                                        1, 2, 3, 4, 5, 1, 1, 1, 1, 1
#>  5 4847                                                    1, 2, 3, 1, 1, 1
#>  6 4848                                              1, 2, 3, 4, 1, 1, 1, 1
#>  7 4849                                                          1, 2, 1, 1
#>  8 4850                                  1, 2, 4, 5, 6, 7, 1, 1, 1, 1, 1, 1
#>  9 4851                                                          1, 2, 1, 1
#> 10 4854                                                          1, 2, 1, 1
#> 11 4855                                        1, 2, 3, 4, 5, 1, 1, 1, 1, 1
#> 12 4857                1, 2, 3, 4, 5, 6, 7, 8, 9, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 13 4858  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 14 4859                                        1, 2, 3, 4, 5, 1, 1, 1, 1, 1
#> 15 4861  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 16 4862                1, 2, 3, 4, 5, 6, 7, 8, 9, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 17 4863                                                          1, 2, 1, 1
#> 18 4864                                                          1, 2, 1, 1
#> 19 4865                                                    1, 2, 3, 1, 1, 1