Pivot data from wide to long.

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. See more details in vignette("pivot"), and see pivot_wider() for the inverse transformation.

pivot_longer(df, cols, names_to = "name", names_prefix = NULL,
  names_sep = NULL, names_pattern = NULL, values_to = "value",
  values_drop_na = FALSE, col_ptypes = list(), spec = NULL)

pivot_longer_spec(df, cols, names_to = "name", values_to = "value",
  names_prefix = NULL, names_sep = NULL, names_pattern = NULL,
  col_ptypes = NULL)

Arguments

df
A data frame to pivot.

cols
Columns to pivot into longer format. This takes a tidyselect specification.

names_to
A string specifying the name of the column to create from the data stored in the column names of df.

Can be a character vector, creating multiple columns, if names_sep or names_pattern is provided.

names_prefix
A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern
If names_to contains multiple values, these arguments control how the column name is broken up.

names_sep takes the same specification as separate(), and can either be a numeric vector (specifying positions to break on), or a single string (specifying a regular expression to split on).

names_pattern takes the same specification as extract(), a regular expression containing matching groups (()).

If these arguments does not give you enough control, use pivot_longer_spec() to create a spec object and process manually as needed.

values_to
A string specifying the name of the column to create from the data stored in cell values. If names_to is a character containing the special .value sentinel, this value will be ignored, and the name of the value column will be derived from part of the existing column names.

values_drop_na
If TRUE, will drop rows that contain only NAs in the value_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in df were created by its structure.

col_ptypes
A list of of column name-prototype pairs.

If not specified, the type of the generated from names_to will be character, and the type of the variables generated from values_to will be the common type of the input columns used to generate them.

spec
Alternatively, instead of providing cols (and names_to and values_to) you can parse a specification data frame. This is useful for more complex pivots because it gives you greater control on how metadata stored in the column names turns into columns in the result.

Must be a data frame containing character .name and .value columns.