Replace missing values

Replace missing values

replace_na(data, replace, ...)

Arguments

data
A data frame or vector.

replace
If data is a data frame, a named list giving the value to replace NA with for each column. If data is a vector, a single value used for replacement.

...
Additional arguments for methods. Currently unused.

See also

Examples

library(dplyr)
df <- tibble(x = c(1, 2, NA), y = c("a", NA, "b"), z = list(1:5, NULL, 10:20))
df %>% replace_na(list(x = 0, y = "unknown"))

#> # A tibble: 3 x 3
#>       x y       z         
#>   <dbl> <chr>   <list>    
#> 1     1 a       <int [5]> 
#> 2     2 unknown <NULL>    
#> 3     0 b       <int [11]>

df %>% mutate(x = replace_na(x, 0))

#> # A tibble: 3 x 3
#>       x y     z         
#>   <dbl> <chr> <list>    
#> 1     1 a     <int [5]> 
#> 2     2 <NA>  <NULL>    
#> 3     0 b     <int [11]>

# NULL are the list-col equivalent of NAs
df %>% replace_na(list(z = list(5)))

#> # A tibble: 3 x 3
#>       x y     z         
#>   <dbl> <chr> <list>    
#> 1     1 a     <int [5]> 
#> 2     2 <NA>  <dbl [1]> 
#> 3    NA b     <int [11]>

df$x %>% replace_na(0)

#> [1] 1 2 0

df$y %>% replace_na("unknown")

#> [1] "a"       "unknown" "b"