Separate one column into multiple columns.

Given either regular expression or a vector of character positions, separate() turns a single character column into multiple columns.

separate(data, col, into, sep = "[^[:alnum:]]+", remove = TRUE,
  convert = FALSE, extra = "warn", fill = "warn", ...)

Arguments

data
A data frame.

col
Bare column name.

into
Names of new variables to create as character vector.

sep
Separator between columns.

If character, is interpreted as a regular expression. The default value is a regular expression that matches any sequence of non-alphanumeric values.

If numeric, interpreted as positions to split at. Positive values start at 1 at the far-left of the string; negative value start at -1 at the far-right of the string. The length of sep should be one less than into.

remove
If TRUE, remove input column from output data frame.

convert
If TRUE, will run type.convert with as.is = TRUE on new columns. This is useful if the component columns are integer, numeric or logical.

extra
If sep is a character vector, this controls what happens when there are too many pieces. There are three valid options:

"warn" (the default): emit a warning and drop extra values.

"drop": drop any extra values without a warning.

"merge": only splits at most length(into) times

fill
If sep is a character vector, this controls what happens when there are not enough pieces. There are three valid options:

"warn" (the default): emit a warning and fill from the right

"right": fill with missing values on the right

"left": fill with missing values on the left

...
Defunct, will be removed in the next version of the package.

See also

unite(), the complement. separate_ for a version that uses regular evaluation and is suitable for programming with.

Examples

library(dplyr)
df <- data.frame(x = c(NA, "a.b", "a.d", "b.c"))
df %>% separate(x, c("A", "B"))

#>      A    B
#> 1 <NA> <NA>
#> 2    a    b
#> 3    a    d
#> 4    b    c

# If every row doesn't split into the same number of pieces, use
# the extra and file arguments to control what happens
df <- data.frame(x = c("a", "a b", "a b c", NA))
df %>% separate(x, c("a", "b"))

#> Warning: Too many values at 1 locations: 3

#> Warning: Too few values at 1 locations: 1

#>      a    b
#> 1    a <NA>
#> 2    a    b
#> 3    a    b
#> 4 <NA> <NA>

# The same behaviour but no warnings
df %>% separate(x, c("a", "b"), extra = "drop", fill = "right")

#>      a    b
#> 1    a <NA>
#> 2    a    b
#> 3    a    b
#> 4 <NA> <NA>

# Another option:
df %>% separate(x, c("a", "b"), extra = "merge", fill = "left")

#>      a    b
#> 1 <NA>    a
#> 2    a    b
#> 3    a  b c
#> 4 <NA> <NA>

# If only want to split specified number of times use extra = "merge"
df <- data.frame(x = c("x: 123", "y: error: 7"))
df %>% separate(x, c("key", "value"), ": ", extra = "merge")

#>   key    value
#> 1   x      123
#> 2   y error: 7