separate_rows {tidyr}	R Documentation

Separate a collapsed column into multiple rows

Description

If a variable contains observations with multiple delimited values, this separates the values and places each one in its own row.

Usage

separate_rows(data, ..., sep = "[^[:alnum:].]+", convert = FALSE)

Arguments

data
A data frame.

...
<tidy-select> Columns to separate across multiple rows

sep
Separator delimiting collapsed values.

convert
If TRUE will automatically run type.convert() on the key column. This is useful if the column types are actually numeric, integer, or logical.

Examples

df <- tibble(
  x = 1:3,
  y = c("a", "d,e,f", "g,h"),
  z = c("1", "2,3,4", "5,6")
)
separate_rows(df, y, z, convert = TRUE)

[Package tidyr version 1.1.4 Index]