tidyr-package {tidyr}	R Documentation

tidyr: Tidy Messy Data

Description

Tools to help to create tidy data, where each column is a variable, each row is an observation, and each cell contains a single value. 'tidyr' contains tools for changing the shape (pivoting) and hierarchy (nesting and 'unnesting') of a dataset, turning deeply nested lists into rectangular data frames ('rectangling'), and extracting values out of string columns. It also includes tools for working with missing values (both implicit and explicit).

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://tidyr.tidyverse.org

https://github.com/tidyverse/tidyr

Report bugs at https://github.com/tidyverse/tidyr/issues

[Package tidyr version 1.1.4 Index]