Unnest a list column.

If you have a list-column, this makes each element of the list its own row. unnest() can handle list-columns that contain atomic vectors, lists, or data frames (but not a mixture of the different types).

unnest(data, ..., .drop = NA, .id = NULL, .sep = NULL,
  .preserve = NULL)

Arguments

data
A data frame.

...
Specification of columns to unnest. Use bare variable names or functions of variables. If omitted, defaults to all list-cols.

.drop
Should additional list columns be dropped? By default, unnest will drop them if unnesting the specified columns requires the rows to be duplicated.

.id
Data frame identifier - if supplied, will create a new column with name .id, giving a unique identifier. This is most useful if the list column is named.

.sep
If non-NULL, the names of unnested data frame columns will combine the name of the original list-col with the names from nested data frame, separated by .sep.

.preserve
Optionally, list-columns to preserve in the output. These will be duplicated in the same way as atomic vectors. This has dplyr::select semantics so you can preserve multiple variables with .preserve = c(x, y) or .preserve = starts_with("list").

Details

If you unnest multiple columns, parallel entries must have the same length or number of rows (if a data frame).

See also

Examples

library(dplyr)
df <- tibble(
  x = 1:3,
  y = c("a", "d,e,f", "g,h")
)
df %>%
  transform(y = strsplit(y, ",")) %>%
  unnest(y)

#>   x y
#> 1 1 a
#> 2 2 d
#> 3 2 e
#> 4 2 f
#> 5 3 g
#> 6 3 h

# Or just
df %>%
  unnest(y = strsplit(y, ","))

#> # A tibble: 6 x 2
#>       x y    
#>   <int> <chr>
#> 1     1 a    
#> 2     2 d    
#> 3     2 e    
#> 4     2 f    
#> 5     3 g    
#> 6     3 h    

# It also works if you have a column that contains other data frames!
df <- tibble(
  x = 1:2,
  y = list(
   tibble(z = 1),
   tibble(z = 3:4)
 )
)
df %>% unnest(y)

#> # A tibble: 3 x 2
#>       x     z
#>   <int> <dbl>
#> 1     1     1
#> 2     2     3
#> 3     2     4

# You can also unnest multiple columns simultaneously
df <- tibble(
 a = list(c("a", "b"), "c"),
 b = list(1:2, 3),
 c = c(11, 22)
)
df %>% unnest(a, b)

#> # A tibble: 3 x 3
#>       c a         b
#>   <dbl> <chr> <dbl>
#> 1    11 a         1
#> 2    11 b         2
#> 3    22 c         3

# If you omit the column names, it'll unnest all list-cols
df %>% unnest()

#> # A tibble: 3 x 3
#>       c a         b
#>   <dbl> <chr> <dbl>
#> 1    11 a         1
#> 2    11 b         2
#> 3    22 c         3

# You can also choose to preserve one or more list-cols
df %>% unnest(a, .preserve = b)

#> # A tibble: 3 x 3
#>       c b         a    
#>   <dbl> <list>    <chr>
#> 1    11 <int [2]> a    
#> 2    11 <int [2]> b    
#> 3    22 <dbl [1]> c    

# Nest and unnest are inverses
df <- data.frame(x = c(1, 1, 2), y = 3:1)
df %>% nest(y)

#>   x data
#> 1 1 3, 2
#> 2 2    1

df %>% nest(y) %>% unnest()

#>   x y
#> 1 1 3
#> 2 1 2
#> 3 2 1

# If you have a named list-column, you may want to supply .id
df <- tibble(
  x = 1:2,
  y = list(a = 1, b = 3:4)
)
unnest(df, .id = "name")

#> # A tibble: 3 x 3
#>       x     y name 
#>   <int> <dbl> <chr>
#> 1     1     1 a    
#> 2     2     3 b    
#> 3     2     4 b