Unnest a list-column of data frames

unnest(), unnest_longer(), and unnest_wider() flatten list-columns into regular columns. unnest() is designed primarily for lists of data frames, where unnest_wider() and unnest_longer() are designed specifically for lists of vectors.

Learn more in vignette("rectangling").

unnest(data, cols, ..., keep_empty = FALSE, ptype = NULL,
  names_sep = NULL, name_repair = "check_unique",
  .drop = "DEPRECATED", .id = "DEPRECATED", .sep = "DEPRECATED",
  .preserve = "DEPRECATED")

unnest_longer(df, cols, value_to = "values", index_to = "index",
  keep_empty = FALSE, names_sep = NULL, name_repair = "check_unique")

unnest_wider(df, cols, names_sep = NULL, name_repair = "check_unique")

Arguments

data
A data frame.

cols
Names of columns to unnest.

If you unnest() multiple columns, parallel entries must compatible sizes, i.e. they're either equal or length 1 (following the standard tidyverse recycling rules).

...
Deprecated: Convert df %>% unnest(x, y) to df %>% unnest(c(x, y)) and df %>% unnest(y = fun(x, y, z)) to df %>% mutate(y = fun(x, y, z)) %>% unnest(y).

keep_empty
By default, elements of col that have size zero will be ommitted from the output. Setting keep_empty = TRUE will ensure that they're preserved

ptype
Optionally, supply a data frame prototype for the output cols, overriding the default that will be guessed from the combination of individual values.

names_sep
If NULL, the default, the names of new columns will come directly from the inner data frame.

If a string, the names of the new columns will be formed by pasting together the outer column name with the inner names, separated by names_sep.

name_repair
Used to check that output data frame has valid names. Must be one of the following options:

"minimal": No name repair or checks, beyond basic existence,

"unique": Make sure names are unique and not empty,

"check_unique": (the default), no name repair, but check they are unique,

"universal": Make the names unique and syntactic

a function: apply custom name repair.

a formula: a purrr-style anonymous function (see rlang::as_function())

See tibble::name-repair for more details on these terms and the strategies used to enforce them.

.drop, .preserve
Deprecated: all list-columns are now preserved; If there are any that you don't want in the output use select() to remove them prior to unnesting.

.id
Deprecated: convert df %>% unnest(x, .id = "id") to df %>% mutate(id = names(x)) %>% unnest(x)).

.sep
Deprecated: use names_sep instead.

df
A data frame.

value_to
Name of column to store vector values.

index_to
A string giving the name of a new column which will contain the inner names of the values. If unnamed, col will instead contain numeric indices.

Unnest variants

The three unnest() functions differ in how they change the shape of the output data frame:

unnest_wider() preserves the rows, but changes the columns.

unnest_longer() preserves the columns, but changes the rows

unnest() can change both rows and columns.

These principles guide their behaviour when they are called with a non-primary data type. For example, if you unnest_wider() a list of data frames, the number of rows must be preserved, so each column is turned into a list column of length one. Or if you unnest_longer() a list of data frame, the number of columns must be preserved so it creates a packed column. I'm not sure how if these behaviours are useful in practice, but they are theoretically pleasing.

See also

Examples

# unnest() is primarily design to work with nested columns, i.e. lists
# of tibbles or data frames
df <- tibble(
  x = 1:3,
  y = list(
    NULL,
    tibble(a = 1, b = 2),
    tibble(a = 1:3, b = 3:1)
  )
)
df %>% unnest(y)

#> # A tibble: 4 x 3
#>       x     a     b
#>   <int> <dbl> <dbl>
#> 1     2     1     2
#> 2     3     1     3
#> 3     3     2     2
#> 4     3     3     1

df %>% unnest(y, keep_empty = TRUE)

#> # A tibble: 5 x 3
#>       x     a     b
#>   <int> <dbl> <dbl>
#> 1     1    NA    NA
#> 2     2     1     2
#> 3     3     1     3
#> 4     3     2     2
#> 5     3     3     1

# You can use unnest_longer() and unnest_wider() with nested dfs,
# although it's not clear how useful the results are. unnest_longer()
# maintains the same number of columns, creating a packed data frame,
# while unnest_wider() maintains the same number of rows, creating
# list-cols of vectors
df %>% unnest_wider(y)

#> # A tibble: 3 x 3
#>       x a         b        
#>   <int> <list>    <list>   
#> 1     1 <NULL>    <NULL>   
#> 2     2 <dbl [1]> <dbl [1]>
#> 3     3 <int [3]> <int [3]>

df %>% unnest_longer(y)

#> # A tibble: 4 x 2
#>       x   y$a    $b
#>   <int> <dbl> <dbl>
#> 1     2     1     2
#> 2     3     1     3
#> 3     3     2     2
#> 4     3     3     1

# Typically, however, you'll use unnest_longer() and _wider() with
# list-cols containing vectors
df <- tibble(
  x = 1:3,
  y = list(NULL, 1:3, 4:5)
)
df %>% unnest_longer(y)

#> # A tibble: 5 x 3
#>       x values index
#>   <int>  <int> <int>
#> 1     2      1     1
#> 2     2      2     2
#> 3     2      3     3
#> 4     3      4     1
#> 5     3      5     2

df %>% unnest_longer(y, keep_empty = TRUE)

#> # A tibble: 6 x 3
#>       x values index
#>   <int>  <int> <int>
#> 1     1     NA    NA
#> 2     2      1     1
#> 3     2      2     2
#> 4     2      3     3
#> 5     3      4     1
#> 6     3      5     2

# Automatically creates names if widening
df %>% unnest_wider(y)

#> New names:
#> * `` -> ...1
#> * `` -> ...2
#> * `` -> ...3

#> New names:
#> * `` -> ...1
#> * `` -> ...2

#> # A tibble: 3 x 4
#>       x  ...1  ...2  ...3
#>   <int> <int> <int> <int>
#> 1     1    NA    NA    NA
#> 2     2     1     2     3
#> 3     3     4     5    NA

# And similarly if the vectors are named
df <- tibble(
  x = 1:2,
  y = list(c(a = 1, b = 2), c(a = 10, b = 11, c = 12))
)
df %>% unnest_wider(y)

#> # A tibble: 2 x 4
#>       x     a     b     c
#>   <int> <dbl> <dbl> <dbl>
#> 1     1     1     2    NA
#> 2     2    10    11    12

df %>% unnest_longer(y)

#> # A tibble: 5 x 3
#>       x values index
#>   <int>  <dbl> <chr>
#> 1     1      1 a    
#> 2     1      2 b    
#> 3     2     10 a    
#> 4     2     11 b    
#> 5     2     12 c    

# You can unnest multiple columns simultaneously
df <- tibble(
 a = list(c("a", "b"), "c"),
 b = list(1:2, 3),
 c = c(11, 22)
)
df %>% unnest(c(a, b))

#> # A tibble: 3 x 3
#>   a         b     c
#>   <chr> <dbl> <dbl>
#> 1 a         1    11
#> 2 b         2    11
#> 3 c         3    22

# Compare with unnesting one column at a time, which generates
# the Cartesian product
df %>% unnest(a) %>% unnest(b)

#> # A tibble: 5 x 3
#>   a         b     c
#>   <chr> <dbl> <dbl>
#> 1 a         1    11
#> 2 a         2    11
#> 3 b         1    11
#> 4 b         2    11
#> 5 c         3    22